feat: move Reference model into accounts and remove refs app
