+++ accounts/models.py
@@ -10,6 +10,10 @@ from django.db.models.signals import m2m_changed
 from django.dispatch import receiver
 from datetime import timedelta
 from django.contrib.contenttypes.models import ContentType
+import hashlib
+from io import BytesIO
+from django.core.files.base import ContentFile
+import qrcode
 
 from integrate.models import Entity, EntityUserManager
 
@@ -168,6 +172,55 @@ class User(Entity, AbstractUser):
     def __str__(self):
         return self.username
 
+
+class Reference(Entity):
+    """Store a piece of reference content which can be text or an image."""
+
+    TEXT = "text"
+    IMAGE = "image"
+    CONTENT_TYPE_CHOICES = [
+        (TEXT, "Text"),
+        (IMAGE, "Image"),
+    ]
+
+    content_type = models.CharField(
+        max_length=5, choices=CONTENT_TYPE_CHOICES, default=TEXT
+    )
+    alt_text = models.CharField("Title / Alt Text", max_length=500)
+    value = models.TextField(blank=True)
+    file = models.FileField(upload_to="refs/", blank=True)
+    image = models.ImageField(upload_to="refs/qr/", blank=True)
+    uses = models.PositiveIntegerField(default=0)
+    method = models.CharField(max_length=50, default="qr")
+    include_in_footer = models.BooleanField(
+        default=False, verbose_name="Include in Footer"
+    )
+    created = models.DateTimeField(auto_now_add=True)
+    author = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name="references",
+        null=True,
+        blank=True,
+    )
+
+    def save(self, *args, **kwargs):
+        if self.method == "qr" and self.value:
+            qr = qrcode.QRCode(box_size=10, border=4)
+            qr.add_data(self.value)
+            qr.make(fit=True)
+            img = qr.make_image(fill_color="black", back_color="white")
+            buffer = BytesIO()
+            img.save(buffer, format="PNG")
+            filename = hashlib.sha256(self.value.encode()).hexdigest()[:16] + ".png"
+            if self.image:
+                self.image.delete(save=False)
+            self.image.save(filename, ContentFile(buffer.getvalue()), save=False)
+        super().save(*args, **kwargs)
+
+    def __str__(self) -> str:  # pragma: no cover - simple representation
+        return self.alt_text
+
 class RFID(Entity):
     """RFID tag that may be assigned to one account."""
 
@@ -229,7 +282,7 @@ class RFID(Entity):
         default=BLACK,
     )
     reference = models.ForeignKey(
-        "refs.Reference",
+        "Reference",
         null=True,
         blank=True,
         on_delete=models.SET_NULL,
