--- refs/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-
-
-class RefsConfig(AppConfig):
-    """Configuration for the Refs application."""
-
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "refs"
-
