--- refs/models.py
@@ -1,100 +1,0 @@
-import hashlib
-from io import BytesIO
-
-from django.conf import settings
-from django.core.files.base import ContentFile
-from django.db import models
-from integrate.models import Entity
-from django.contrib.contenttypes.fields import GenericForeignKey
-from django.contrib.contenttypes.models import ContentType
-from django.utils.text import slugify
-from django.utils.translation import gettext_lazy as _
-import qrcode
-
-
-class Reference(Entity):
-    """Store a piece of reference content which can be text or an image."""
-
-    TEXT = "text"
-    IMAGE = "image"
-    CONTENT_TYPE_CHOICES = [
-        (TEXT, "Text"),
-        (IMAGE, "Image"),
-    ]
-
-    content_type = models.CharField(
-        max_length=5, choices=CONTENT_TYPE_CHOICES, default=TEXT
-    )
-    alt_text = models.CharField("Title / Alt Text", max_length=500)
-    value = models.TextField(blank=True)
-    file = models.FileField(upload_to="refs/", blank=True)
-    image = models.ImageField(upload_to="refs/qr/", blank=True)
-    uses = models.PositiveIntegerField(default=0)
-    method = models.CharField(max_length=50, default="qr")
-    include_in_footer = models.BooleanField(
-        default=False, verbose_name="Include in Footer"
-    )
-    created = models.DateTimeField(auto_now_add=True)
-    author = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        on_delete=models.CASCADE,
-        related_name="references",
-        null=True,
-        blank=True,
-    )
-
-    def save(self, *args, **kwargs):
-        if self.method == "qr" and self.value:
-            qr = qrcode.QRCode(box_size=10, border=4)
-            qr.add_data(self.value)
-            qr.make(fit=True)
-            img = qr.make_image(fill_color="black", back_color="white")
-            buffer = BytesIO()
-            img.save(buffer, format="PNG")
-            filename = hashlib.sha256(self.value.encode()).hexdigest()[:16] + ".png"
-            if self.image:
-                self.image.delete(save=False)
-            self.image.save(filename, ContentFile(buffer.getvalue()), save=False)
-        super().save(*args, **kwargs)
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return self.alt_text
-
-class Tag(Entity):
-    """A label that can be applied to any model instance."""
-
-    name = models.CharField(max_length=50, unique=True)
-    slug = models.SlugField(max_length=50, unique=True, blank=True)
-
-    class Meta:
-        ordering = ["name"]
-        verbose_name = _("tag")
-        verbose_name_plural = _("tags")
-
-    def save(self, *args, **kwargs):
-        if not self.slug:
-            self.slug = slugify(self.name)
-        super().save(*args, **kwargs)
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return self.name
-
-
-class TaggedItem(Entity):
-    """Link a Tag to any model instance."""
-
-    tag = models.ForeignKey(Tag, on_delete=models.CASCADE, related_name="tagged_items")
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
-    content_object = GenericForeignKey("content_type", "object_id")
-
-    class Meta:
-        indexes = [models.Index(fields=["content_type", "object_id"])]
-        unique_together = ("tag", "content_type", "object_id")
-        verbose_name = _("tagged item")
-        verbose_name_plural = _("tagged items")
-
-    def __str__(self) -> str:  # pragma: no cover - simple representation
-        return f"{self.tag} -> {self.content_object}"
-
-
