--- refs/templates/refs/footer.html
@@ -1,40 +1,0 @@
-{% load ref_tags %}
-<footer class="container mt-4 pt-4">
-  <div class="row align-items-center">
-    <div class="col-lg-3 text-center text-lg-start mb-3 mb-lg-0">
-      {% current_page_qr 100 %}
-    </div>
-    <div class="col-lg-9">
-      <div class="row justify-content-center">
-        {% for ref in footer_refs %}
-          <div class="col-auto">
-            <a href="{{ ref.value }}">{{ ref.alt_text }}</a>
-          </div>
-        {% endfor %}
-      </div>
-    </div>
-  </div>
-  {% if request.user.is_staff and admin_links %}
-  <div class="row justify-content-center text-secondary small mt-2">
-    {% for name, url in admin_links %}
-    <div class="col-auto"><a class="link-secondary" href="{{ url }}">{{ name }}</a></div>
-    {% endfor %}
-  </div>
-  {% endif %}
-  {% if version or revision %}
-  <div class="row">
-    <div class="col text-end text-muted small">
-      {% if version %}ver {{ version }}{% endif %}{% if revision %} rev {{ revision }}{% endif %}
-    </div>
-  </div>
-  {% endif %}
-</footer>
-<style>
-  html[data-bs-theme='light'] footer {
-    background-color: #e9ecef;
-  }
-  html[data-bs-theme='dark'] footer {
-    background-color: #1a1e21;
-  }
-</style>
-
