--- refs/templates/refs/landing.html
@@ -1,19 +1,0 @@
-{% extends "website/base.html" %}
-{% load i18n %}
-
-{% block title %}{% trans "Reference Generator" %}{% endblock %}
-
-{% block content %}
-<h1>{% trans "Reference Generator" %}</h1>
-<form method="get" class="mb-3">
-  <div class="input-group">
-    <input type="text" class="form-control" name="data" placeholder="Enter text" value="{{ request.GET.data }}" required>
-    <button class="btn btn-primary" type="submit">{% trans "Generate" %}</button>
-  </div>
-</form>
-{% if img_src %}
-<div class="text-center">
-  <img src="{{ img_src }}" alt="QR code" class="img-thumbnail">
-</div>
-{% endif %}
-{% endblock %}
