+++ release/utils.py
@@ -110,7 +110,6 @@ def _write_pyproject(package: Package, version: str, requirements: list[str]) ->
                     "config",
                     "nodes",
                     "ocpp",
-                    "refs",
                     "website",
                     "release",
                     "integrate",
