+++ website/fixtures/gway_box.json
@@ -16,10 +16,6 @@
   },
   {
     "model": "website.application",
-    "fields": {"name": "refs"}
-  },
-  {
-    "model": "website.application",
     "fields": {"name": "msg"}
   },
   {
@@ -46,15 +42,6 @@
       "site": ["192.168.129.10"],
       "application": ["rfid"],
       "path": "/rfid/",
-      "is_default": false
-    }
-  },
-  {
-    "model": "website.siteapplication",
-    "fields": {
-      "site": ["192.168.129.10"],
-      "application": ["refs"],
-      "path": "/refs/",
       "is_default": false
     }
   }
