+++ website/templates/website/base.html
@@ -1,4 +1,4 @@
-{% load i18n ref_tags %}
+{% load i18n %}
 <!doctype html>
 <html lang="{{ LANGUAGE_CODE }}" data-bs-theme="light">
   <head>
@@ -104,7 +104,6 @@
       </nav>
       {% block content %}{% endblock %}
     </div>
-    {% render_footer %}
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script>
       function setTheme(theme) {
